shareddata: Fix race if mult. peers join same time
The issue happens when multiple clients try to join a computation at the
same time. One of the clients gets the lock in the kernel to create the
shared data file (using `open(O_EXCL)`). Now, this client (the leader)
is responsible for initializing the shared data header, and the other
clients must wait for the initialization to complete.

The current test for completion of initialization was inadequate; it
relied on testing the size of the shared data file. The race condition
here is that client that gets the lock to create the shared data file
would set the size (using `truncate`) and be preempted before it got a
chance to finish the initialization of the header. Meanwhile, the other
clients would race ahead, check the version string in the shared data
header, and assert out.

The fix here is to let the clients wait until the leader has finished
initializing the shared data.