sysvipc: Fix a race with two clients creating shm with IPC_PRIVATE
There was a subtle race condition in our `shmget()` wrapper with two
clients trying to simultaneously create an shm region with `IPC_PRIVATE`
and `IPC_EXCL`.

The code relies on the `VIRTUAL_TO_REAL_SHM_KEY` API, which, in turn,
uses the DMTCP shared data area to share and assign unique keys for SysV
shm regions. The API returns -1 if the key does not exist in the DMTCP
mapping. Now, if one client raced ahead and managed to create the region,
it'd end up populating the DMTCP shared area with a "IPC_PRIVATE ->
realKey" mapping. Later, when the other client would try to create its
own shared memory region, the `VIRTUAL_TO_REAL_SHM_KEY` API would return
the key corresponding the earlier created entry and the `shmget(IPC_EXCL)`
call would fail with an `EEXIST` error.

The fix here is to detect this special case of `IPC_PRIVATE` and use
a process's real pid to create the shared memory region. This also
preserves the semantics of `IPC_PRIVATE`.