Include barrier name in DmtcpMessage.
Sometimes the process might get killed after writing DmtcpMessage but
before writing "extra-data" on the coordinator socket. Currently, the
coordinator is not robust enough to detect it. The fix is to send the
barrier name as part of DmtcpMessage.

Since the DmtcpMessage struct is of static size, we are imposing a
limit of 64 characters for the barrier names.