Fix to sporadic crash on restarted application
TLSInfo_GetTidOffset() sometimes calculates the wrong tid offset.  When the function identifies the value is invalid, it returns a statically calculated one.  However, the function still saves the invalid one for use later.  This can cause subsequent mischief and segfaults in libc calls in the restarted function.

This change does not fix the original problem (wrong tid offset), but has the function at least retain the more sensible backup value when it finds the one it calculates is invalid.